aispeak/requirements.txt
aiocoap==0.3
alabaster==0.7.10
Babel==2.4.0
backports.weakref==1.0rc1
baidu-aip==1.6.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto3==1.4.4
botocore==1.5.85
bs4==0.0.1
cbor==1.0.0
certifi==2017.4.17
cffi==1.10.0
cfscrape==1.8.0
chardet==3.0.4
click==6.7
CoAPthon==4.0.2
colorama==0.3.9
CommonMark==0.7.3
decorator==4.0.10
Django==1.10.6
docutils==0.13.1
entrypoints==0.2.2
Flask==0.12.1
Flask-SQLAlchemy==2.2
future==0.16.0
gTTS==1.2.0
gTTS-token==1.1.1
hkdf==0.0.3
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itchat==1.3.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
keyring==10.3.2
LinkHeader==0.4.3
Markdown==2.2.0
MarkupSafe==1.0
mistune==0.7.3
mqtt-benchmark==0.1.0
Naked==0.1.31
nbconvert==5.0.0
nbformat==4.2.0
nltk==3.2.4
notebook==4.3.0
numpy==1.13.0
olefile==0.44
paho-mqtt==1.2
pandocfilters==1.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.9
protobuf==3.3.0
pycparser==2.17
PyExecJS==1.4.0
pygame==1.9.3
Pygments==2.2.0
pymongo==3.4.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
pytz==2017.2
pywin32-ctypes==0.1.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.17.3
robotframework==3.0.2
s3transfer==0.1.10
shellescape==3.4.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinx-argparse==0.2.1
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.10
termcolor==1.1.0
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==1.2.6
win-unicode-console==0.5
